feat: Migrate /development-api/* into /api/*
