+++ src/css/_includes/development-api.scss
@@ -21,6 +21,13 @@
 .api-block-header.request {
   background-color: var(--gray-dark);
   color: var(--light);
+  display: flex;
+  overflow-x: auto;
+  white-space: nowrap;
+
+  :last-child {
+    padding-right: 10px;
+  }
 }
 
 .api-block-header.response {
