+++ src/templates/apiPage.tsx
@@ -4,7 +4,7 @@ import Prism from "prismjs";
 
 import BasePage from "~src/components/basePage";
 import SmartLink from "~src/components/smartLink";
-import DevelopmentApiSidebar from "~src/components/developmentApiSidebar";
+import ApiSidebar from "~src/components/apiSidebar";
 
 import {
   OpenAPI,
@@ -100,7 +100,7 @@ export default props => {
   }, []);
 
   return (
-    <BasePage sidebar={<DevelopmentApiSidebar />} {...props}>
+    <BasePage sidebar={<ApiSidebar />} {...props}>
       <div className="row">
         <div className="col-6">
           {data.summary && <p>{data.summary}</p>}
@@ -152,7 +152,7 @@ export default props => {
               <div>
                 <div>
                   {"You need to "}
-                  <SmartLink to={"/development-api/auth"}>
+                  <SmartLink to={"/api/auth"}>
                     authenticate via bearer auth token.
                   </SmartLink>
                 </div>
@@ -175,7 +175,7 @@ export default props => {
               <span className="api-request-block-verb">
                 {data.method.toUpperCase()}
               </span>{" "}
-              {data.apiPath}
+              <span>{data.apiPath}</span>
             </div>
             <pre className="api-block-example request">
               {apiExample.join(" \\\n")}
